Introduced TemplatedErrorHandler
- Creates a TemplatedErrorHandler based on the previous ErrorHandler; this one:
  - renames `handleError()` to `handleErrorResponse()`
  - adds two *protected* methods, `handleError()` and `handleException()` to
    which `handleErrorResponse()` now delegates; these are extension points.
- Renames ErrorHandler to WhoopsErrorHandler, and has it extend
  TemplatedErrorHandler. It now overrides `handleException()` to provide Whoops
  output.
- Created a factory for the TemplatedErrorHandler.
- Renamed the ErrorHandlerFactory to WhoopsErrorHandlerFactory.
